+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## [3.0.1](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v3.0.0...v3.0.1) (2023-01-24)
+
+
+### Bug Fixes
+
+* stop propagation on mousedown event ([45f79db](https://gitlab.com/dd3tech/dd360/dd360-components/commit/45f79dba9e0e7252061a42be1f8eb404dc7b1534))
+
+# [3.0.0](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v2.5.4...v3.0.0) (2023-01-24)
+
+
+### Bug Fixes
+
+* add improvements to the code: useCallback and usePortal prop and new tests ([614fca7](https://gitlab.com/dd3tech/dd360/dd360-components/commit/614fca726802576678905708782edd39a99527a7))
+
+
+### BREAKING CHANGES
+
+* 🧨 the use of the portal, add more tests to achieve 100 coverage
+
+## [2.5.4](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v2.5.3...v2.5.4) (2023-01-23)
+
+
+### Bug Fixes
+
+* with certain languages set in the configuration, the format was displayed incorrectly ([beafe13](https://gitlab.com/dd3tech/dd360/dd360-components/commit/beafe13d7dc4f97b5620b5e2d2712995cfca68aa))
+
 ## [2.5.3](https://gitlab.com/dd3tech/dd360/dd360-components/compare/v2.5.2...v2.5.3) (2023-01-23)